create reusable component
